functions/yt-extract/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
requests
python-dateutil
